requirements.txt
asgiref==3.4.1
#backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
bleach==4.1.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
Django==3.2.10
django-contrib-comments==2.1.0
filebrowser-safe==1.0.0
grappelli-safe==1.0.0
idna==3.3
Mezzanine==5.0.0
oauthlib==3.1.1
packaging==21.3
Pillow==8.4.0
pyparsing==3.0.6
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
webencodings==0.5.1
gunicorn==20.1.0
psycopg2-binary==2.8.6
google-cloud-secret-manager==2.4.0
google-auth==1.30.0
#django-storages[google]==1.11.1
google-cloud-storage==1.10.0  
django-storages==1.7.1
django-environ==0.4.5
